[svn-r4049] This was a big mistake in the proxy implementation that was completely ignoring the return statuses of the processes and returning success to the launcher every time. This commit should fix that part and some other minor bugs in the process cleanup.

+++ src/pm/hydra/pm/central/proxy_cb.c
@@ -60,7 +60,8 @@ HYD_Status HYD_Proxy_listen_cb(int fd, HYD_Event_t events)
 
         if (cmd == KILLALL_PROCS) {     /* Got the killall command */
             for (i = 0; i < HYD_Proxy_params.proc_count; i++)
-                kill(HYD_Proxy_params.pid[i], SIGKILL);
+                if (HYD_Proxy_params.pid[i] != -1)
+                    kill(HYD_Proxy_params.pid[i], SIGKILL);
 
             status = HYD_DMX_Deregister_fd(fd);
             if (status != HYD_SUCCESS) {
@@ -190,32 +191,4 @@ HYD_Status HYD_Proxy_stdin_cb(int fd, HYD_Event_t events)
 
   fn_fail:
     goto fn_exit;
-}
-
-
-void HYD_Proxy_signal_cb(int signal)
-{
-    int i;
-
-    HYDU_FUNC_ENTER();
-
-    if (signal == SIGINT || signal == SIGQUIT || signal == SIGTERM
-#if defined SIGSTOP
-        || signal == SIGSTOP
-#endif /* SIGSTOP */
-#if defined SIGCONT
-        || signal == SIGCONT
-#endif /* SIGSTOP */
-) {
-        /* There's nothing we can do with the return value for now. */
-        for (i = 0; i < HYD_Proxy_params.proc_count; i++)
-            kill(HYD_Proxy_params.pid[i], SIGKILL);
-        exit(-1);
-    }
-    else {
-        /* Ignore other signals for now */
-    }
-
-    HYDU_FUNC_EXIT();
-    return;
 }
